put storage and server behind featuregate

+++ pkg/controllers/core/catalog_controller.go
@@ -82,26 +82,28 @@ func (r *CatalogReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ct
 	if err := r.Client.Get(ctx, req.NamespacedName, &existingCatsrc); err != nil {
 		return ctrl.Result{}, client.IgnoreNotFound(err)
 	}
-	if existingCatsrc.DeletionTimestamp.IsZero() {
-		if !controllerutil.ContainsFinalizer(&existingCatsrc, fbcDeletionFinalizer) {
-			controllerutil.AddFinalizer(&existingCatsrc, fbcDeletionFinalizer)
-			if err := r.Update(ctx, &existingCatsrc); err != nil {
-				return ctrl.Result{}, err
+	if features.CatalogdFeatureGate.Enabled(features.HTTPServer) {
+		if existingCatsrc.DeletionTimestamp.IsZero() {
+			if !controllerutil.ContainsFinalizer(&existingCatsrc, fbcDeletionFinalizer) {
+				controllerutil.AddFinalizer(&existingCatsrc, fbcDeletionFinalizer)
+				if err := r.Update(ctx, &existingCatsrc); err != nil {
+					return ctrl.Result{}, err
+				}
 			}
+		} else {
+			if controllerutil.ContainsFinalizer(&existingCatsrc, fbcDeletionFinalizer) {
+				err := r.Storage.Delete(existingCatsrc.Name)
+				if err != nil {
+					return ctrl.Result{}, err
+				}
+				// remove our finalizer from the list and update it.
+				controllerutil.RemoveFinalizer(&existingCatsrc, fbcDeletionFinalizer)
+				if err := r.Update(ctx, &existingCatsrc); err != nil {
+					return ctrl.Result{}, err
+				}
+			}
+			return ctrl.Result{}, nil
 		}
-	} else {
-		if controllerutil.ContainsFinalizer(&existingCatsrc, fbcDeletionFinalizer) {
-			err := r.Storage.Delete(existingCatsrc.Name)
-			if err != nil {
-				return ctrl.Result{}, err
-			}
-			// remove our finalizer from the list and update it.
-			controllerutil.RemoveFinalizer(&existingCatsrc, fbcDeletionFinalizer)
-			if err := r.Update(ctx, &existingCatsrc); err != nil {
-				return ctrl.Result{}, err
-			}
-		}
-		return ctrl.Result{}, nil
 	}
 	reconciledCatsrc := existingCatsrc.DeepCopy()
 	res, reconcileErr := r.reconcile(ctx, reconciledCatsrc)
@@ -168,13 +170,14 @@ func (r *CatalogReconciler) reconcile(ctx context.Context, catalog *v1alpha1.Cat
 		//   of the unpacking steps.
 
 		fbc, err := declcfg.LoadFS(unpackResult.FS)
-		if err := r.Storage.Store(catalog.Name, fbc); err != nil {
-			return ctrl.Result{}, updateStatusUnpackFailing(&catalog.Status, fmt.Errorf("error storing fbc: %v", err))
-		}
 		if err != nil {
 			return ctrl.Result{}, updateStatusUnpackFailing(&catalog.Status, fmt.Errorf("load FBC from filesystem: %v", err))
 		}
-
+		if features.CatalogdFeatureGate.Enabled(features.HTTPServer) {
+			if err := r.Storage.Store(catalog.Name, fbc); err != nil {
+				return ctrl.Result{}, updateStatusUnpackFailing(&catalog.Status, fmt.Errorf("error storing fbc: %v", err))
+			}
+		}
 		if features.CatalogdFeatureGate.Enabled(features.PackagesBundleMetadataAPIs) {
 			if err := r.syncPackages(ctx, fbc, catalog); err != nil {
 				return ctrl.Result{}, updateStatusUnpackFailing(&catalog.Status, fmt.Errorf("create package objects: %v", err))
